requirements.txt
django == 3.1.2
django-admin-interface
django-crispy-forms
pillow
django-rest-framework
requests
django-paypal
paypal-checkout-serversdk